Making sure guest authors are enabled before calling related functions to avoid Fatal.

+++ co-authors-plus.php
@@ -940,15 +940,17 @@ class CoAuthors_Plus {
 			wp_delete_term( $delete_user->user_login, $this->coauthor_taxonomy );
 		}
 
-		// Get the deleted user data by user id.
-		$user_data = get_user_by( 'id', $delete_id );
+		if ( $this->is_guest_authors_enabled() ) {
+			// Get the deleted user data by user id.
+			$user_data = get_user_by( 'id', $delete_id );
+		
+			// Get the associated user.
+			$associated_user = $this->guest_authors->get_guest_author_by( 'linked_account', $user_data->data->user_login );
 
-		// Get the associated user.
-		$associated_user = $this->guest_authors->get_guest_author_by( 'linked_account', $user_data->data->user_login );
-
-		if ( isset( $associated_user->ID ) ) {
-			// Delete associated guest user.
-			$this->guest_authors->delete( $associated_user->ID );
+			if ( isset( $associated_user->ID ) ) {
+				// Delete associated guest user.
+				$this->guest_authors->delete( $associated_user->ID );
+			}
 		}
 	}
